backend/requirements.txt
landingai-ade>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.2
Pillow>=10.0.0
